Notebooks/Solutions/Day2/requirements.txt
gradio
pandas
langchain
langchain-community
langchain-openai
faiss-cpu
python-dotenv
openai
